+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog for shamir
 
+## 0.2
+
+* Improve performance of Newton's algorithm
+* Update documentation
+
 ## 0.1
 
 * Implement Shamir secret sharing using Lagrange interpolation